// src/models/GetInfo.ts
// Generated by https://quicktype.io



export class ArtistInfo {
    artist: ArtistInfoArtist = new ArtistInfoArtist();
}

export class ArtistInfoArtist {
    name:       string = "";
    mbid:       string = "";
    url:        string = "";
    image:      Image[] = [];
    streamable: string = "";
    ontour:     string = "";
    stats:      Stats = new Stats();
    similar:    Similar = new Similar();
    tags:       Tags = new Tags();
    bio:        Bio = new Bio();
}

export class Bio {
    links:     Links = new Links();
    published: string = "";
    summary:   string = "";
    content:   string= "";
}

export class Links {
    link: Link = new Link();
}

export class Link {
    "#text": string = "";
    rel:     string= "";
    href:    string= "";
}

export interface Image {
    "#text": string;
    size:    string;
}

export class Similar {
    artist: ArtistElement[] = [];
}

export interface ArtistElement {
    name:  string;
    url:   string;
    image: Image[];
}

export class Stats {
    listeners: string = "";
    playcount: string = "";
}

export class Tags {
    tag: Tag[] = [];
}

export interface Tag {
    name: string;
    url:  string;
}